Testing "framework" extended with pattern matching
Tests can now be written for openscad activity that is naively
examined after execution for the existance or absence of text.

Tests that end in `.yscad` are treated as sh(1) or bash(1) compliant
scripts and sourced. They're expected to define at a minimum a `SCAD`
variable assignment, within which is the OpenSCAD content you wish to test.
The `SCAD` variable is dumped into a temporary file within the `test/` subdirectory,
and then has `run-test.sh` execute it as normal. The temporary file is cleaned up
afterwards.
Optionally, the `.yscad` file can define a `MATCH_TEXT` and `NOMATCH_TEXT`
variable: if these are set, the output of the OpenSCAD run is examined accordingly.
If NOMATCH_TEXT is set and that content is present in the test run output, the test
is considered failed; if MATCH_TEXT is set and that content is *not* present in the test
run output, the test is considered failed.

Original, simple testing - a basic execution and assumption of any content output
to STDOUT or STDERR as a failure - is unchanged.