Conditional logging by test, per-level
New logging modules and functions added, a la `log_debug_if()` and similarly named module/functions.
The first argument is a boolean test, and can be specified as a straight `true` or `false`, or by
putting a statement to be evaluated (eg, `1 > 0`, or `evaluate_this_function_for_truth(arg)`).
If the test evaluates as `true`, the message is to be logged (provided `LOG_LEVEL` is set to
an appropriate level for that called). If the test evaluates as `false`, the message is not logged
regardless of `LOG_LEVEL`.

`log_debug_if()` (and its ilk) will return the evaluated test when called as a function.

`log_fatal_if()` will also return the evaluated test, but the nature of `log_fatal_if()` precludes the
use of that value (since the execution of the libfile will already have stopped).