added log_*_unless(), and log_*_assign()
log_WHATEVER_unless() is the boolean opposite of log_WHATEVER_if():
given a test, if the test is false, log the message. if the test is
true, do nothing. Essentially, "log this message unless this test is true".

log_WHATEVER_assign() logs at whatever level that somethign is being assigned,
and unconditionally returns the value given. There's no log_fatal_assign(),
because what would be the point?